// src/lib/api/generated/model/publicPageSchema.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * COMPASS API
 * The canonical backend API for COMPASS clients.
 * OpenAPI spec version: 1.0.0
 */

export interface PublicPageSchema {
  audience: string;
  body_html: string;
  id: string;
  page_key: string;
  published_at?: string | null;
  status: string;
  summary: string;
  title: string;
}